Allow oidc config without well_known.
Until now it was not possible to configure the oidc config without a well known endpoint and was limited to a keycloak specific endpint.
With this change one can now choose between providing the complete well known endpoint or configuring all the necessary endpoints.

Also the change introduces an API change of the config object of OIDC.