Allow custom injectables to enrich the opa payload.
Injectables can now be added to the opa config. There is an interface class for injectables.
The injectables allow to enrich the token sent to opa with custom information.

The GraphQLAnalyzer can now be used in this way. There is also an injectable for this kind o enrichment provided.
Furthermore the tests show how this can be used.