Adapt auth interface incl. request.
For SAML auth it is required to have the complete request provided (form data) instead of only
the connection. Also it makes sense to enforce async for the auth method.
The change does not need a major bump, because request inherits from connection and
given implementations should be compatible with the change.