Allow openapi endpoints.
Fastapi brings openapi docs by default.
The endpoints (/docs, /redoc and openapi.json) should not be protected by this middleware.
If they need to be hidden this can be done by changing the apps settings instead:
https://fastapi.tiangolo.com/advanced/conditional-openapi/?h=openapi#conditional-openapi-from-settings-and-env-vars

The fix does not yet modify the docs so that the endpoints appear protected.