Bump GitPython to fix RCE and update packages.
To fix remote code execution using GitPython (https://github.com/advisories/GHSA-hcpj-qp55-gfph) this version has to be bumped.

Also in the process of fixing that issue this commit updates packages
and due to that some tests needed a fix regarding the http method in
use (get instead of post was used while using the json parameter for
data).